fix: fix session serialization for PUSH with websocket
When PUSH transport is websocket, the request associate to atmosphere is a stub implementation
that always return false for 'isRequestedSessionIdValid()' preventing PushSessionTracker to
serialize the session.
In addition, accessing the request from a background thread may throw IllegalStateException if
the servlet container is recycling request objects (e.g. Tomcat facade).
With this change, the session validiy is checked against the SessionListener that is already
observing session lifecycle events.

Fixes #119