fix: shutdown SessionSerializer executor service
When running the debug tool, a new SessionSerializer instance is created for every request,
but once the serialization/deserialization process finish the internal executor service
is not stopped.
This change stop the executor service in the ContextClosedEvent, and it makese
the debug tool destroy the temporary session serializer instance.
In addition some cleanup are performed eagerly.