fix: prevent usage of VaadinRequest in access
VaadinRequest thread local is not available during a VaadinSession.access execution.
VaadinRequest is used to access the HTTP session wrapper, that can be obtained through
VaadinSession. The most important usage is however getting the update version header,
but the lookup can be done outside the access block.
This change also removes ClusterSupport registration in CurrentInstance becuase it
seems not to be used anywhere and furhtermore the thread local is set only on the
thread that calls serviceInit and never cleaned up.

Fixes #148