fix: ensure SessionTrackerFilter is executed before Spring Security
Currently, Spring Security filter chain is executed before SessionTrackerFilter
restores the HTTP session, causing Vaadin security checks to fail even if the
user information are available in the serialized session.
This change fixes the order of the SessionTrackerFilter, so that it is always
executed before the Spring Security filter chain.

Fixes #170